Swap the two parents of ScriptEngineLoader
My understanding of the reason Nashorn started failing to load
in Java 9 has been somewhat mistaken since 6cd539b. It is not
so much that the Maven-supplied class loader never delegates to
Java's loader. The key change came with the Java 9 module system,
and the way the ServiceLoader works for services in named modules.

The former (and current, for unnamed modules) algorithm involved
searching for META-INF/services/* resources to find the providers
of services. For named modules, the bindings are computed from the
module-info files, and the javadoc for ServiceLoader.load provides:

"Step 1: Locate providers in named modules.

Service providers are located in all named modules of the class loader
or to any class loader reachable via parent delegation."

That turns out to be literally true. The ServiceLoader is not guided
by whatever delegation policy a class loader actually implements by
the way it overrides ClassLoader's default search order. To identify
reachable modules, ServiceLoader only knows to follow a ClassLoader's
actual parent link, nothing else.

That means this two-parent classloader idea wasn't wrong, but if it's
to find providers in Java-supplied modules, it needs its parents
swapped so the Java application class loader is the real one and the
loader supplied by Maven is the 'extra' one.

With that delegation order, it could present some risk of class version
conflicts, if the Maven-supplied loader has defined some classes that
should supersede those from the application loader. That could be fixed
by overriding other ClassLoader methods to change the standard delegation
strategy. But with the narrow, targeted use of this loader only with the
ScriptEngineManager, the risk may be low enough.

It is fortunate that, for now, Maven itself and Maven plugins are not
heavy users of named modules. This headache could recur if it ever
becomes necessary to load named-module services from *both* parents
of this loader. At that point, it could become necessary to explore
whether the comments in ServiceLoader.load about ModuleLayer searching
would offer any solution.