It's the platform loader, not the application one
I had been relying on [1] in believing that Nashorn had been moved
onto the application class path, but it turns out to be found with
no trouble by the platform class loader. If the Maven developers
were concerned about stability if delegating to the application
loader, that concern should be reduced by delegating just to the
platform loader instead.

[1] https://issues.apache.org/jira/browse/MANTRUN-200?focusedCommentId=16138712&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-16138712