Tackle argument passing via pg_ctl and on Windows
Wasn't expecting to have to bite this bullet quite so soon,
but discovering that pg_ctl also constructs command lines in
egregiously injectable ways kind of moved the schedule up.

Introduce two methods that 'transform' a ProcessBuilder (by
rewriting its 'command' argument list).

forWindowsCRuntime assumes the starting argument list is what
the target program is intended to get as its arguments, and that
it is a C program that will parse them apart using the C/C++
rules covered in [1], and rewrites them with awareness of both
those parsing rules and what the Java runtime does to the arguments
before the process is launched (from painstaking examination of
the ProcessImpl source).

asPgCtlInvocation assumes the starting argument list is what would
start a server by invoking postgres directly, and transforms it to
one that would invoke it via pg_ctl. Because of the way pg_ctl
handles options, this transformation needs to be platform-aware
and apply quoting for Bourne shell or Windows CMD as appropriate.
Even that only covers getting the args safely from pg_ctl to
postgres. Getting them safely from Java to pg_ctl in the first place
requires (on Windows) following this transformation with
forWindowsCRuntime.

[1] http://www.daviddeley.com/autohotkey/parameters/parameters.htm#WINARGV