Maven really hates a dependency version of LATEST
It emits large clamorous warnings about how it threatens the
stability of the build and someday they might not even support
building such malformed projects, you project malformer, you.

Of course LATEST was only used here in the optional downloading
of pgjdbc-ng, which is only done for automated testing, and if
automated testing were to uncover a later version of pgjdbc-ng
where something stopped working, that would be useful to know.

Changed LATEST to [0.8.4,) which is just as open-ended, just as
threatening to the stability of the build, and in that sense just
as malformed, but happily accepted by Maven without a peep.