A step or two further
The javadoc for pljava/ doesn't fail now. It was the module-path for
its pljava-api dependency. I've hardcoded it here, but really, I am
pretty sure Maven has a method "org.postgresql:pljava-api is one of
my dependencies, what is the path to that jar?" That method (or a
convenience wrapper for it) should be exposed to the script engine
so this can be done right. In pljava-examples/ there is also a
dependency on the pljava-api jar; that should also come from Maven.

There is still something very puzzling about pljava/ though.
It is documenting only the mbeans package, even though it is
being passed --show-module-contents all and --show-packages all,
and it is complaining about "unmappable character ... for encoding
US-ASCII" even though it is being passed both -encoding UTF-8
and -docencoding UTF-8. It is almost as if it is ignoring some
arguments. I have not determined why.

For pljava-examples it was complaining about non-ASCII characters
and I added the -encoding and -docencoding and it stopped. So they
aren't ignored there. Puzzling.

To see the generated Maven reports, run mvn site site:stage
and then use a web browser to open target/staging/index.html.
You can click one of the links under Modules (pljava-api,
pljava, pljava-examples), and then under Project Reports click
Documentation Report. It's blank. The script needs to be able
to get a reference to the Doxia Sink (the script engine used to
get a 'report' entry referencing 'this'; probably it still should,
and getSink() can be called on that. The script should output
to the sink at least a heading and a link to apidocs/index.html.