An AutoCloseable with exception-type parameter
Java's AutoCloseable does not have a parameter for the thrown
exception type. It is declared Exception, and while subinterfaces
are "strongly encouraged" to narrow that declaration, that's of
no use to the compiler except in code that explicitly uses the
subinterface.

In passing move the Trivial interface inside a class (arbitrarily,
the Closing class just added here), only so it can be made private;
it has a default method that performs an unchecked cast, and should
only be extended by other interfaces in this compilation unit.