Add a test class for Checked
Even compile-time only 'tests' with no runtime behavior are enough
to flush out some typos and usability issues.

The several overloaded in() methods seem to require the compiler
to infer more than it always can. Give those distinct names.

Much is not yet tested. Everything won't be ('everything' would
include the Cartesian product of everything with a use() method
and the ten flavors of in...() method ... fuhgeddaboudit.