Fix bug in javadoc generation
TL,DR; override isExternalReport to return true
There was a flaw in the logic of the ReportScriptingMojo. The
index.html under apidocs is created by the javadoc tool. Before this
commit, the mojo created an index.html of its own in the same place
overwriting the one created by the javadoc tool. Deleting the snippet
creating this file is not enough because maven would automatically
generate a file with the project styling but empty contents of its own.
This would again override the index.html created by javadoc tool.

Maven obtains the name of the html to generate from getOutputName
method of the report. This file name is automatically linked to the
project-reports.html page so changing the name would spill its own
problems. However, the solution is fairly simple. Maven Reporting API
provides a isExternalReport method which can be overridden to turn off
this behaviour.