Create ReportScript interface for extensibility
Some parameters like getOutputName and so on were hardcoded inside the
ReportScriptingMojo. While this was enough to meet the current needs of
generating javadoc, it could be problematic with other future potential
use cases. Therefore, a ReportScript interface has been created with
some relevant methods that are subset of AbstractMavenReport class.

The script can extend the interface and override only the values it
wants to. For the remaining, the defaults will be used. This new
instance must then be passed to the ReportScriptingMojo using
setReportScript.