Use Invocable for ReportScript
ReportScript is an interface that the scripting engine (provided
it implements Invocable) can automagically generate an implementation
of, as long as the script block that has been evaluated contained
appropriate function definitions.

It is even possible to give the interface default methods, which
the script block can override, but doesn't have to.

The interface methods end up needing the mojo as a parameter
(so the default interface methods have a way to refer to it), and
of course that means the script functions need to accept that
parameter as well. That's ok; the script needed a reference to
the mojo anyway. It used to be bound in the script engine with the
name 'report'; now it's a function parameter. The world still spins.

If the interface default methods to return the same defaults
implemented by AbstractMavenReport, they just have to call back on
some new, non-public methods on the mojo, which do nothing but
call the corresponding superclass methods.