Adjust the javadoc scripts to match
It turns out getOutputName should return a relative path
from the report output directory, ending in the basename of
the file that Maven should generate a link to. So if javadoc
will be told to write into reportOutputDirectory/apidocs/
then getOutputName should return apidocs/index and Maven will
generate a relative link to apidocs/index.html.