Improve pljava and pljava-api javadoc links
When documenting modular code, javadoc generates an index.html
with nothing much in it but a link to the module-summary.html
for the module. We can eliminate one click for the person
reading the docs by having getOutputName return the relative
path to the module-summary file instead, so Maven generates
a link directly to it.

I am already liking this better than the maven-javadoc-plugin.

For pljava/ though, it still seems to be ignoring --show-packages all,
and for the life of me I still don't know why.