Ensure plugin works in Java 15 for reports too
Profiles that add dependencies to pljava-pgxs can't be in
the parent pom, as they would cause failure in a first-time
build where that plugin isn't built yet. So they have to be
duplicated in each submodule that uses pljava-pgxs, and there
are more of those now, thanks to its use for javadocs. Not
delightful, but bearable.

In profiles that add dependencies bar and baz to plugin foo,
it isn't necessary to specify the version of foo; that would
just become a value duplicated in many places and probably
overlooked for updates.

The profile in the parent pom that adds the graaljs dependencies
to maven-antrun-plugin still needs to be there, as long as we are
still using that plugin to run any scripts. One day if all those
scripts have been migrated to use the new plugin, that will
become unnecessary. That profile is also responsible for setting
the polyglot.js.nashorn-compat system property to true, without
which the Graal engine behaves differently from the Nashorn one.

That turns out not to be enough, because properties-maven-plugin
doesn't run in the 'site' lifecycle, but the report scripts still
need the property. So a bit of hardcoding in PGXSUtils will also
set it, if it isn't set and this is Java 15 or later. Also not
delightful, but it seems to be necessary. That means that the
setting of that property via profile in the parent pom could also
become unnecessary if all uses of script in antrun get migrated
to the new plugin.

In passing, fix an older apparent logic bug in PGXSUtils and put
a comment back on the line where it belonged.