Work on getting the javadocs right
The javadoc for pljava/ was messing up in surprising ways,
complaining about non-ASCII source characters even though
-encoding was specified, and ignoring the --show-module-contents
and --show-packages specifications. The ignoring of -encoding was,
of all things, the result of including --release 9. Apparently
there is some bug where using the --release option clobbers an
-encoding option.

The --show-* trouble was pilot error: avoid using the legacy
one-stop-shopping options like -private/-package/-protected/
-public in combination with the newer -show-* options, because
the former can clobber the latter.

To aid comprehension and troubleshooting, group the javadoc
options the same way they are in the tool's documentation:
first the ones that match javac options, then the core javadoc
options, then the ones that are interpreted by the standard
doclet.

Phase out interpolated strings like ${project.name} in the
JavaScript. It was convenient for prototyping and pleasant to
look at, but of course Maven knows nothing about JavaScript
syntax and it would make a mess if an interpolated value had
special characters in it. The same values are available via
access to the 'report' parameter.

Because the --module-source-path mod=path syntax is used, and
that appeared in javadoc 12, override canGenerateReport and
return false if the JVM is earlier. There may be other ways of
getting the javadoc to generate in 9 through 11, but the extra
complexity might be worse than just saying "you need 12 or later
to build the javadocs".