A few remaining options the Maven plugin supplied
The computation of 'bottom' here depends on the values not needing
any HTML escaping (but that's pretty much how the Maven plugin did it,
too). Some other day might feel like a good day to improve it.

There seems to be no need for -charset. A meta tag gets emitted anyway
with the right encoding (probably picked up from -docencoding; that
would be the way to do it).