Try to get links to other docs working
When requiring javadoc >= 12, trying to link back to Oracle's
Java 10 docs will not work, because javadoc 12 assumes the tree
has module names as the top directory level, and Oracle's
pre-Java-12 trees do not. So, construct the 'jdklink' uri
for 12 rather than 10.

This -linkoffline to the pljava-api docs 'works' in the sense
that javadoc finds them and generates links. However, it
generates links that don't work. The reason: well, javadoc's
documentation for -linkoffline clearly says url1 can be
absolute or relative, and that, when relative, it "should be
the relative path from the destination directory (specified
with the -d option) ...". In other words, the documentation
contract commits javadoc to generate links with as many ..
elements added as necessary to get from the referring page's
depth in the tree back to the -d root, and then resolve the
relative url1 from there. And this clearly documented, contracted
behavior is exactly what it does not do. So if it places a
relative external link into a page sitting four package levels
down from the -d root, it puts the supplied url1 there with no
adjustment for depth, and the link is guaranteed to be broken.

No usable workaround has occurred to me yet.

Also need to do something about the basedir.resolve(srcroot).
There are two Path methods, resolve(Path) and resolve(String),
and srcroot is a Path, and Nashorn thinks that's perfectly clear
to figure out which method to call, but Graal thinks it's just
too hard to figure out.