Copy javadoc tweaks to other submodules
Still lacking a workaround for javadoc's botching of relative
url1 in -linkoffline. Because Path.resolve(Path.toString())
happens to be equivalent to Path.resolve(Path) (at least when
the argument Path is from the default filesystem provider, as
will be the case here), adding the toString() makes an easy
workaround for graalvm/graaljs#254 ... though it is likely that
more headaches will arise from that bug, and workarounds may
not always be so easy.

Also change a couple of links in Markdown documentation to go
directly to pljava-api's module summary. Links to index.html
get redirected there if browser JavaScript is enabled, but if
JavaScript filtering is in place, there's an inelegant stop
at a mostly blank page with a link.