(Try to) link -api from -examples if Java >= 15
Not until Java 15 does javadoc learn it can link to modular docs
from nonmodular code (it still warns, as if that isn't a perfectly
reasonable thing to do while not all code in the world has been
modularized, but at least it does the job). Earlier versions log
errors rather than warnings, and don't emit the links, so just
skip that -linkoffline option for them.

The links still don't *work*, for reasons amply covered in 82efab96,
but at least they are links. To make them work, against javadoc's
best efforts, is going to be a headache.