Introduce a RelativizingFileManager
This is a workaround for javadoc's -linkoffline behavior, where
following the documented instruction to supply a URL relative to
the -d destination results in broken links, because javadoc forgets
to add the right number of ../ components when writing files in
subdirectories, as it would have to for that rule to work. It also
fails to expand {@docRoot} in a -linkoffline URL, which would
otherwise be a simpler workaround.

This file manager simply interposes on output HTML files as they
are written, and rewrites URLs containing RELDOTS to replace that
token with the appropriate number of ../ components, exactly as
{@docRoot} would be replaced if it were honored.

A supplied -linkoffline URL must have its own ../ components,
if any, placed before RELDOTS and not after, because javadoc
normalizes the URL, such that if RELDOTS were followed by
any ../ component, both would disappear, and this file manager
would find nothing to rewrite.