Do Repeat Yourself
If you pass your own file manager to DocumentationTool's getTask,
it won't find any of your files. This even if the file manager
you supply is a forwarding wrapper around the exact file manager
that getStandardFileManager() gave you.

The reason turns out to be that the location options you pass
to the DocumentationTool (-classpath, -sourcepath, etc.) do not
get passed on to your file manager unless it's of an actual subclass
of the internal, unextendable file manager class that you got
from getStandardFileManager.[1] A forwarding wrapper isn't one of
those, so those options don't get propagated to it, and therefore
you must pass it the same location-defining options that you're
already passing to the DocumentationTool itself.

Who ever said DRY had to stand for *Don't* Repeat Yourself?

This only seems to be needed for the location-defining options
that are in the 'standard options' (inherited from javac) group.
Locations specific to the doclet (like -d) do get passed to
the file manager without extra attention here, by the doclet I guess.

[1] http://hg.openjdk.java.net/jdk9/jdk9/langtools/file/65bfdabaab9c/src/jdk.javadoc/share/classes/jdk/javadoc/internal/tool/Start.java#l540