Extra .. needed when linking from modular code
The RelativizingFileManager has to check whether the output
location is module-oriented, and add another ../ if so, to
account for the module-name directory level at the top of
the tree.

For some reason, javadoc in pljava/ will complain if the -d
option isn't passed to the file manager. In pljava-examples/
it wasn't being passed, and nothing complained. Anyway, moving
-d up to the start of args in both places for consistency.