Only supply -sourcetab 4 for examples
The option was added on speculation that it might improve
the formatting of some generated documentation, by matching the
project's conventional tab width. However, the option implicitly
turns on -linksource, to generate linked HTMLified versions of
the source code, which I'm not sure is necessary (GitHub does a
fair job of that).

It is not clear whether -sourcetab has any effect anywhere else
but the HTMLified sources. But there is not any -nolinksource
option to reverse its implicit -linksource, so simply remove it
again.

Leave it for -examples, where perusing the source is the point.