Get classpath/modulepath segregation working
The pljava-examples javadoc was working, which sounds good,
but was actually weird, because it was only passing -classpath
to javadoc, and buildPaths() should have been detecting that
pljava-api is modular, and putting it on the module path instead,
where javadoc would not have been able to find it.

That was a small matter in buildPaths(): the endsWith() method
on a Path does not mean what endsWith() on a String means. On
a Path, it tests for an ending sequence of intact path elements.
If you want to match just the ending /portion/ of the last path
element against some /characters/, it has to be a String first.

With that working, of course the -examples javadoc fails, which
could be fixed two ways: either smash paths['classpath'] and
paths['modulepath'] back into a single string and call it -classpath
(which is back to what was happening before, and works because the
examples code isn't modular, so even the dependencies that are
modules don't have to be treated as such), or, pass the class path
and the module path separately, and name org.postgresql.pljava in
an --add-modules option (which is needed because the examples code,
being ignorant of modules, has no way of saying "hey, I use that
module").

Done here the second way, as it seems more in the spirit of the
module system.