Handle the empty-argument case for Windows
There will probably never be an empty argument in any expected
invocation of initdb or postgres from here, but make sure the case
is not left unhandled in forWindowsCRuntime anyway, in case the code
gets reused elsewhere.

Kartik confirmed empirically that "" works, and that the Java runtime
doesn't catch the case itself.