Have pljava-pgxs use itself to make its javadoc
This requires, again, a profile declaring Graal's JavaScript
engine as a dependency of the plugin for Java >= 15.

In this case, though, this IS the plugin, so a simpler profile
can be used that directly declares the dependencies (unlike the
profiles added in other subprojects, which had to use the more
elaborate syntax to declare dependencies of a plugin they used).

A convenient consequence of a profile in the plugin's own pom,
declaring straight dependencies, is that the profiles added in
other subprojects' poms earlier, to add the same dependencies,
are no longer needed. I never liked having so many copies of those.

This amounts to the pljava-pgxs plugin making a commitment that
JavaScript will always be an available language for scripting.
And that is pretty much the intent.

In passing, remove three declared dependencies that seem unnecessary
to building the plugin. If they need to be added again as more of
its functionality is developed, so be it, but for now they seem
unneeded. If we ever want to make its javadoc link offline to the
corresponding Maven docs, it will seem less daunting without such
a long list of dependencies.