Add semiFlattenDiagnostics
The flattenDiagnostics method relies on the behavior of
SQLException.iterator(), which will produce one flattened stream
of SQLExceptions or SQLWarnings and their causes, if any.

In many cases, it will be preferable only to flatten the
chain represented by getNextException or getNextWarning, leaving
causes to be retrieved from the exception or warning they caused.
That's what semiFlattenDiagnostics does. It is now the default
behavior for qp with no 'flattener' argument; a new qp that takes
a flattener argument is supplied to preserve the choice.