Add a rudimentary DFA for checking a result stream
The caller can supply a list of lambdas representing numbered states of
a DFA, where the first in the list is considered state 1 and the initial
state. Repeatedly, the current state will be applied to the current item
of input, and may return the number of the next state, true to indicate
a successful match, or false for inability to proceed. When moving to
a new state, the current input item may be consumed, or retained and
examined again in the new state.

Also add a method peek(Object), which will print an object largely the
same way qp would, but intended to be passed to Stream.peek to create a
stream that will visibly print its members while also being passed to
dfa() for matching. Unlike qp, peek declares no checked exceptions (as
the Stream API requires), and will print only the metadata of a
ResultSet, avoiding closing it or disturbing its cursor, so it can still
be examined in the DFA.