Use Node.dfa to confirm expected results
... and arrange for jshell to exit with nonzero status as appropriate.

In jshell 10 and later, /exit can take a value, but in 9 it can't,
so System.exit must be used (and works ok in this setting with
--execution local).

An uncaught throwable (1) doesn't result in a nonzero exit, or any
exit at all, and (2) gets printed more concisely than it would be
by printStackTrace. So in the catch block, just record that it
happened, and rethrow it.