Skip Windows JDK9 builds
jshell 9 did not yet understand the file name - to mean stdin,
read non-interactively.

On Linux, that just means it complains about the file name -
and then reads from stdin anyway, printing all prompts and input
as if it were interactive, which looks goofy but otherwise works.

On Windows, it seems to complain about the file name - and then
sit forever waiting for nonexistent console input. A JDK 9 Windows
test would probably require writing the jshell script into a temp
file and putting that name on the jshell command line.