Allow different granularity file/process times
With OS X and Java 9 there could be a frequent hang because
the modification time on postmaster.pid appeared truncated to
seconds, and would look earlier than the process start time,
which wasn't. As the time check is just a heuristic, add a second
to the file time before comparing.

Caught with some logic to detect cycles in the wait_for_pid_file
routine, which will be left in place as a hedge against future
similar puzzles.