Try to prove PGROOT and similar vars unnecessary
The key is to find the right pg_config for the PostgreSQL version
to be built and tested against, and be sure to pass that on the
command line to mvn for building and to the package jar for installing
and testing. If any of those other environment manipulations are needed,
there's a bug somewhere.

Also run the chosen pg_config, so its output is in the log in case of
questions about what was actually being built.

Ended up breaking out the appveyor build_script to a separate BAT file,
because if you put a multiline CMD block in appveyor.yml it gets passed
to CMD line-by-line and breaks ... unlike a multiline PowerShell block,
which is handled as one would expect ... but PowerShell messes up what
external commands write to standard error, so would be awkward to use
here.