Minor polishing, skip updating homebrew
... assuming whatever homebrew cache is initially on the VM image
probably is up-to-date enough? That should at least slim the log
by the large lists of updated formulae. (Maybe that is an indication
that what's on the VM image is not up-to-date enough. Have to see.)

Eliminating the homebrew auto update also eliminates the auto-upgrade
it was doing to dozens of packages we have no interest in, cutting that
part of the build from over 9 minutes down to half a minute or so.