Maybe uninitialized
The BEGIN/END_NATIVE family of macros conceal 'if' statements:
if certain conditions are pending, the underlying check 'throws'
an exception and returns false. Ordinarily, a thrown exception
would mean no code after the 'if' ever executes, and there would
be no concern about using a variable whose initialization was
contained in the if block.

In JNI, though, an exception isn't really thrown until the
function returns. It just has to skip the actions inside the
block and return something, anything, as directly as possible.
At that point, the JVM throws the exception and no use of the
return value is made, but the C compiler still sees that the
function could return an uninitialized value in some circumstances.
Simply initializing it, even to an arbitrary value, quiets
the warning.