Fix MinGW build by removing -DBUILDING_DLL
Windows has a particular global symbol filtering mechanism. The
-DBUILDING_DLL flag is present in CPPFLAGS used by PostgreSQL to build
its dlls. In context of the PostgreSQL build, the flag is used to
mark the symbols that should be exposed to other object files. However,
PL/Java is a PostgreSQL extension. It only needs to import the symbols
from PostgreSQL header files and not export those. What happens is that
-DBUILDING_DLL causes PGDLLIMPORT marker to expand to
__declspec(dllexport) instead of the usual __declspec(dllimport). This
leads to a failure during linking of object files. Removing this flag
from CPPFLAGS fixes the issue.