Return MavenReportException from executeReport
If an exception occurs during execution of ReportScriptingMojo, it
should be thrown so that Maven knows about it and stops the execution
immediately. However, Nashorn and Graal wrap exceptions thrown during
execution in different ways. Nashorn wraps the original exception twice
whereas Graal consumes the original exception and always throws a
Polyglot exception wrapped in ScriptException. Therefore, directly
throwing an exception can obscure the actual cause. Hence, we return the
 actual exception, if one occurred, from the method. If execution
 completes successfully, null is returned instead.