Use more casual term 'state machine' than 'dfa'
There was already another one added in 4e001ae, and there
the less-jargon-y term 'state machine' was used; for consistency,
use that in Node also. Anyway, because the states can be
arbitrary lambdas, it doesn't have to be strictly a DFA; it
would be simple to implement a pushdown automaton, for example
(or just call another stateMachine from a state). It is easy to
foresee tests that might do such things (to check a sequence
of results by descending into the individual ResultSets and into
individual rows of those, for example). So stateMachine is
just a better name.