Accept possible flags in pg_config --cc
In some cases, pg_config can sneak flags into the reported CC value.
For example, a simple source build of PostgreSQL 12 in my Linux
environment results in "gcc -std=gnu99" as the value of CC.

So it seems it has to be treated as a list, then its first element
used as the executable, and any others as flags.