Avoid MSVC TZ workaround in PG 10.2 and later
In 10.2, session_timezone gained a PGDLLIMPORT marking, making
the workaround unnecessary. This typo'd conditional caused it
to be used in later 10.x versions anyway.

Tempting to blame the change in PG_VERSION_NUM format with PG 10
(in 958fe54), but the value here was wrong for the old format too.

Addresses #297. Apparently the workaround wasn't a perfect one,
either.