Add GoalScript to execute script using Invocable
Nashorn wraps exceptions thrown during Javascript execution whereas
Graal consumes them and throws an altogether new exception. In order to
correctly handle exceptions thrown during execution of the script, we
return the exception after wrapping it inside MojoExecutionException.

As it is not possible to return a value from the script directly, we use
the GoalScript interface to execute the script from Java and obtain the
return value. A non null value indicates unsuccessful completion.