Put a lid on permissions when appropriate
PL/Java has long allowed a CREATE FUNCTION to directly name some
class/method that isn't in a PL/Java-managed jar, such as something in
the Java runtime API itself. It's a little weird, but an established
usage: just about everybody's first PL/Java blog post tests out a
CREATE FUNCTION ... AS 'java.lang.System.getProperty' (which is even
included in pljava-examples).

Under policy-based permissions, the Java runtime itself is unrestricted,
which of course is not the desired behavior. There is no need to forbid
using CREATE FUNCTION to name non-PL/Java-managed code, but it needs to
be invoked with a restrictive AccessControlContext, here one that
represents "nobody special". That is, it will enjoy only those permissions
the policy blanket-grants to all domains.

There may still be room to fine-tune the criteria for when to apply this
lid. To all classes not in sqlj.install_jar() jars? To only Java runtime
classes? If an admin drops foo.jar onto the classpath and wants to grant
it permissions in the policy, should this lid stand in the way?