Bring Iterator-based SRFs into the scheme
The Invocable for such an SRF will return another Invocable.
That's what will be called through the iterate/close steps of
SRF evaluation; its AccessControlContext will be the same one
that was selected for the SRF itself.

Still to do: the same trick for ResultSetProvider-based SRFs
(and for ResultSetHandle-based ones, which just need an addition
in the MethodHandle tree to instantiate a ResultSetPicker).

Completed, this will obviate the need for Type_hasNextSRF/
nextSRF/closeSRF to be "virtual methods" of Type; the choice
of behavior will just be made up front when building the
method handle.