Cover transaction and savepoint listeners
These will have an AccessControlContext saved when they are registered
and used when they are invoked. The EntryPoints.Invocable class is
used here for convenience, though invocation is done directly, without
going through any EntryPoints method. The saved AccessControlContext
is going to include pljava.internal's protection domain anyway
(without a lot of work to avoid it); using an EntryPoints method
wouldn't help.

A listener that wants to do something that is permitted to its own jar
but not to PL/Java will have to accept the inconvenience of using
doPrivileged.

In passing, adds a Checked.BiConsumer, as it may be useful elsewhere.
(The four-argument consumer needed in SubXactListener is just declared
there as a private interface.)