Add pljava.policy_urls GUC.
Allows setting the pljava.policy URL (and any others as desired)
at chosen places in the policy.url.n sequence begun in the systemwide
java.security file.

While it is not yet (with these patches) possible to grant permissions
based on the code signers of a jar, or the PostgreSQL role under which
the code is executing, it is at least possible to have multiple policy
files and use ALTER ROLE ... SET pljava.policy_urls ...; to select
a specific policy based on the authenticated role at the time of
connection.

Of course only a superuser can set this GUC.