Add linux-ppc64le to builds for Travis CI
The language: java runner environment isn't able to install the same
selection of JDKs on ppc64le as on amd64, so switch instead to the
language: minimal runner, and explicitly get Java from adoptopenjdk.

That can be done with the install-jdk.sh script, which already exists
in /usr/local/bin on and amd64 linux runner ... but not on ppc64le
(or osx, for that matter). So, get that script from GitHub (at a
specific SHA after reviewing it) when it isn't already installed.
(It might be just as well to download it unconditionally, and never
wonder what version the runner has.)

The install-jdk.sh script needs a patch to compute the right
JAVA_HOME value on osx.

In constructing the URL for downloading from adoptopenjdk, the
Travis arch value 'amd64' must be 'x64', and the Travis OS 'osx'
must be 'mac'.

Naturally, the minimal runner doesn't have Maven preinstalled
either. It's not a bad thing to have to fetch Maven; it is good
to confirm the build works back to 3.5.2, as the documentation
claims.

The minimal runner doesn't have PostgreSQL installed, which
doesn't change much, except travis_install_postgresql.sh needs
to be undeterred if 'service postgresql stop' fails. (I don't
know what that chmod was doing there; builds seem to work
without it.)

For Java 9, the --cacerts option has to be given to the
install-jdk.sh script, or Maven won't be able to download
dependencies because it won't be able to verify connections
to repository hosts.

Use the matrix now to generate linux builds for amd64 and ppc64le
with various PostgreSQL versions, Java feature versions, JVM
implementations (hotspot, openj9) and Maven versions. There isn't
support for osx-ppc64le, so remove osx from the matrix, and just
use include: to add some osx builds manually.

The env: section in the matrix is also kind of manually populated
for a plausible subset of possible builds. The full Cartesian
product of {arch}x{pg version}x{java version}x{hotspot,openj9}x
{maven version} would be a test of Travis's generosity.

It is good to have added OpenJ9 into the matrix, and discovered
a segfault. So comment that back out of the matrix for now.