getObject(,Class<?>) to get a primitive array
The C implementations for all of the primitive array types already
had coerceDatum methods producing Java primitive arrays; I'm just
not sure it was ever reachable before. The boxed arrays normally
produced are what the method inherited from Array.c makes.

So change a few lines in Type.c to allow it to happen, and it happens.