Allow primitive array in updateObject()
Once again, the code's been there all along in the C implementations,
with some Java code in SingleRowWriter saying "nope, can't do that."
Allow it to happen, and it happens.

The NullPointerException if the supplied (boxed) array type contains
nulls isn't fixed yet. I did not notice earlier that Byte[], alone
among the boxed primitive array types, does not incur this NPE. It
does incur the value corruption discussed on the mailing list. But
that does not happen with byte[]. So there is clearly something
distinct about how Byte[] is being handled.